10 Deploy MCP Server on EC2/pyproject.toml
[project]
name = "mcp-mastery-claude-and-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.1.0",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.116.1",
    "fastmcp>=2.12.3",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.29",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-ollama>=0.3.8",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "matplotlib>=3.10.6",
    "mcp-use[search]>=1.3.10",
    "mcp[cli]>=1.14.0",
    "numpy>=2.3.3",
    "openai>=1.108.1",
    "pandas>=2.3.2",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "scipy>=1.16.2",
    "streamlit>=1.49.1",
]
